moved parts to new SO class
