+++ Editor/SOVariant.cs
@@ -1,6 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Sirenix.Serialization;
 using UnityEditor;
 using UnityEngine;
+using Object = UnityEngine.Object;
 
 namespace Giezi.Tools
 {
@@ -11,7 +16,229 @@ namespace Giezi.Tools
         public AssetImporter _import;
         public List<string> _overridden;
         public List<string> _otherSerializationBackend;
-        public List<CheckBoxAttribute> _checkBoxAttributes;
         public List<string> _children;
+
+        public bool SetParent(T parent)
+        {
+            if(_target is null)
+                return false;
+        
+            if (parent)
+            {
+                if (parent.GetType() != _target.GetType())
+                {
+                    Debug.Log("Only equal types can be selected as parent");
+                    return false;
+                }
+
+                if (AssetDatabase.GetAssetPath(parent) == AssetDatabase.GetAssetPath(_target))
+                {
+                    Debug.Log("You can't select the same object as parent");
+                    return false;
+                }
+            }
+
+            string targetGUID = AssetDatabase.AssetPathToGUID(AssetDatabase.GetAssetPath(_target));
+            if (_parent)
+                RemoveFromChildrenData(AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(_parent)), targetGUID);
+            AddToChildrenData(AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(parent)), targetGUID);
+        
+            this._parent = parent;
+            SaveData(new List<string>());
+            _overridden = null;
+            return true;
+        }
+
+        public void LoadData(Object targetObject)
+        {
+            try
+            {
+         
+                _target = (T) targetObject;
+
+                string path = AssetDatabase.GetAssetPath(targetObject);
+                _import = AssetImporter.GetAtPath(path);
+            }
+            catch (Exception e)
+            {
+                return;
+            }
+            
+            if(_target is null || _import is null)
+                return;
+                
+            try
+            {
+                string data = _import.userData;
+                var extractedData = ExtractData(data);
+                _parent = extractedData.Item2;
+                _overridden = extractedData.Item3 ?? new List<string>();
+                _children = extractedData.Item4 ?? new List<string>();
+            }
+            catch (Exception e)
+            {
+                _parent = null;
+                _overridden = new List<string>();
+                _children = new List<string>();
+            }
+        }
+
+        private Tuple<string, T, List<string>, List<string>> ExtractData(string data)
+        {
+            string[] datas = data.Split('*');
+            if (datas.Length != 3)
+                return new Tuple<string, T, List<string>, List<string>>(null, null, null, null);
+
+            byte[] parentDataStream = datas[0].Split(',').ToList().Select(source => byte.Parse(source)).ToArray();
+            string parentGUID = SerializationUtility.DeserializeValue<string>(parentDataStream, DataFormat.Binary);
+
+            var parent = AssetDatabase.LoadAssetAtPath<T>(AssetDatabase.GUIDToAssetPath(parentGUID));
+
+            byte[] overridesDataStream = datas[1].Split(',').ToList().Select(source => byte.Parse(source)).ToArray();
+            var overridden = SerializationUtility.DeserializeValue<List<string>>(overridesDataStream, DataFormat.Binary);
+
+            var children = DeserializeChildrenData(datas[2]);
+
+            return new Tuple<string, T, List<string>, List<string>>(parentGUID, parent, overridden, children);
+        }
+
+        private List<string> DeserializeChildrenData(string data)
+        {
+            byte[] childrenDataStream = data.Split(',').ToList().Select(source => byte.Parse(source)).ToArray();
+            var children = SerializationUtility.DeserializeValue<List<string>>(childrenDataStream, DataFormat.Binary);
+            return children;
+        }
+
+        public void SaveData(List<string> overriddenMembers)
+        {
+            if(_import is null)
+                return;
+        
+            string overridesData = SerializeOverrideData(overriddenMembers);
+
+            string parentData = SerializeParentData(_parent);
+
+            PropagateValuesToChildren(_target, AssetDatabase.AssetPathToGUID(AssetDatabase.GetAssetPath(_target)), ref _children, _import);
+        
+            string childrenData = SerializeChildrenData(_children);
+
+            string data = parentData + "*" + overridesData + "*" + childrenData;
+
+            _import.userData = data;
+        
+            EditorUtility.SetDirty(_target);
+            AssetDatabase.SaveAssets();
+        }
+
+        private string SerializeChildrenData(List<string> children)
+        {
+            return string.Join(",",
+                SerializationUtility.SerializeValue<List<string>>(children, DataFormat.Binary));
+        }
+
+        private string SerializeParentData(T parent)
+        {
+            return string.Join(",",
+                SerializationUtility.SerializeValue<string>(AssetDatabase.GUIDFromAssetPath(AssetDatabase.GetAssetPath(parent)).ToString(), DataFormat.Binary));
+        }
+
+        private string SerializeOverrideData(List<string> overrides)
+        {
+            if (_otherSerializationBackend != null && _otherSerializationBackend.Count > 0 && overrides.All(s => s != _otherSerializationBackend.First()))
+                overrides.AddRange(_otherSerializationBackend);
+
+            return string.Join(",",
+                SerializationUtility.SerializeValue<List<string>>(overrides, DataFormat.Binary));
+        }
+
+        private void PropagateValuesToChildren(T target, string targetGUID, ref List<string> children, AssetImporter importer)
+        {
+            bool childrenUpdated = false;
+        
+            foreach (string child in new List<string>(children))
+            {
+                string childPath = AssetDatabase.GUIDToAssetPath(child);
+                AssetImporter childImporter = AssetImporter.GetAtPath(childPath);
+                if(childImporter == null)
+                {
+                    children.Remove(child);
+                    childrenUpdated = true;
+                    continue;
+                }
+            
+                var extractedData = ExtractData(childImporter.userData);
+            
+                if(extractedData.Item1 == null || extractedData.Item2 == null || extractedData.Item3 == null || extractedData.Item4 == null)
+                {
+                    children.Remove(child);
+                    childrenUpdated = true;
+                    continue;
+                }
+            
+                string parentGUID = extractedData.Item1;
+
+                if (parentGUID != targetGUID)
+                {
+                    children.Remove(child);
+                    childrenUpdated = true;
+                    continue;
+                }
+            
+                List<string> overridden = extractedData.Item3;
+                T childObject = AssetDatabase.LoadAssetAtPath<T>(childPath);
+
+                foreach (FieldInfo fieldInfo in FieldInfoHelper.GetAllFields(target.GetType()))
+                {
+                    if(overridden.Contains(fieldInfo.Name))
+                        continue;
+                    object value = FieldInfoHelper.GetFieldRecursively(target.GetType(), fieldInfo.Name).GetValue(target);
+                    FieldInfoHelper.GetFieldRecursively(childObject.GetType(), fieldInfo.Name).SetValue(childObject, value);
+                }
+                EditorUtility.SetDirty(childObject);
+
+
+                List<string> newChildrenList = extractedData.Item4;
+                if(newChildrenList.Count > 0)
+                    PropagateValuesToChildren(childObject, AssetDatabase.AssetPathToGUID(childPath), ref newChildrenList, importer);
+            }
+
+            if (childrenUpdated)
+            {
+                WriteOnlyChildrenData(importer, children);
+            }
+        }
+
+        public void AddToChildrenData(AssetImporter importer, string newChild)
+        {
+            string[] data = importer.userData.Split('*');
+            if (data.Length != 3)
+            {
+                WriteOnlyChildrenData(importer, new List<string>(){newChild});
+                return;
+            }
+
+            List<string> children = DeserializeChildrenData(data[2]);
+            children.Add(newChild);
+            WriteOnlyChildrenData(importer, children);
+        }
+
+        public void RemoveFromChildrenData(AssetImporter importer, string oldChild)
+        {
+            string[] data = importer.userData.Split('*');
+            if(data.Length != 3)
+                return;
+            List<string> children = DeserializeChildrenData(data[2]);
+            children.Remove(oldChild);
+            WriteOnlyChildrenData(importer, children);
+        }
+
+        private void WriteOnlyChildrenData(AssetImporter importer, List<string> children)
+        {
+            string[] data = importer.userData.Split('*');
+            if(data.Length == 3)
+                importer.userData = data[0] + "*" + data[1] + "*" + SerializeChildrenData(children);
+            else
+                importer.userData = SerializeParentData(null) + "*" + SerializeOverrideData(new List<string>()) + "*" + SerializeChildrenData(children);
+        }
     }
 }
